server/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
passlib[argon2]
argon2-cffi
python-jose[cryptography]
python-dotenv
